Fix Signal Flares Admin Log/JMP (#3453)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

Fix issues when jumping to signal flares from the admin niche log. It'd
still be 0,0,0 if the user still has it in hand. Worse, this appears to
crash client when jumping to it -- but that's going to be a different
fix because it probably affects more..


# Changelog
:cl:
fix: Fixed incorrect coordinates in Admin Niche log for signal flares.
/:cl: