Fix Force Dropship Admin Verb (#3514)
# About the pull request

This PR fixes the force dropship admin verb that could be used to hijack
a dropship. Now it will not runtime, nor require the admin to be
adjacent to the console.

~For a non-crash, I would have it just open the shuttle manipulator, but
that is a verb on the admin datum instead of the client~ - this verb
also doesn't perform any admin permissions checking currently... (it
probably should).

# Explain why it's good for the game

Admin verbs should work, otherwise they should be eliminated.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![hijack](https://github.com/cmss13-devs/cmss13/assets/76988376/3e982699-2ec4-4edd-a3ef-ebe5461656a6)


</details>

# Changelog
:cl: Drathek
admin: Fixed the force dropship verb (now known as Force Hijack)
requiring the admin to be adjacent to the console and runtiming.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>