Overwatch Signals for QDEL: Less client disconnection? (#3477)
# About the pull request

This PR is sort of a stab in the dark to fix #3444 by making marine and
xeno overwatching register a signal for qdeletion to end the overwatch
faster. The actual source of this issue is apparently that the
multithreaded map tick beta has an issue with crashing clients, so the
changes made in this PR don't need to be permanent if the effect isn't
as desirable when the beta is fixed.

Locally testing has not been able to reproduce the problem, but I can
confirm I do not get the momentary blank screen when overwatching
something that gets deleted.

# Explain why it's good for the game

Hopefully no more (or less) client disconnections when overwatching
something that gets deleted.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![qdel](https://github.com/cmss13-devs/cmss13/assets/76988376/2d9af325-3959-40ae-a36c-1fc8b2d950cf)

</details>

# Changelog
:cl: Drathek
fix: Clients should hopefully avoid disconnection when observing someone
that is deleted.
/:cl: