Hand and feet now cannot get IB (#3545)
# About the pull request

This PR makes it impossible to get IB in your hands and feet.

# Explain why it's good for the game

Incentivizing the tradeoffs for armored versus unarmored regions when it
comes to choosing a limb to target is important. This further tilts it
towards armored areas which should help promote varied preference.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
balance: Hand and feet now cannot get IB
/:cl: