Burrowers unable to burrow while inside a non-burrow area (#3559)
# About the pull request

This PR makes it so burrowers can no longer burrow while inside a
non-burrow area. Currently, they can burrow but not TUNNEL into a
non-burrow area.

# Explain why it's good for the game

Awkward and niche interaction that is causing issues. You really
shouldn't be able to hide while burrowed most of the way up on a
dropship. If you win the fight and get off the dropship you get to reap
the benefits go wild.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
balance: Burrowers are now unable to burrow while inside a non-burrow
area
/:cl: