Find internal bleeding via Health Scanner (#3561)
# About the pull request

This PR allows finding internal bleeding via the health scanner.

# Explain why it's good for the game

While I think the check status way to do this is cool (and the analyzer
is kind of lame overall(different conversation)), it's just not very
intuitive and so here we are.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
add: You can now find internal bleeding via Health Scanner
/:cl: