Adaptive clientside zoom for pixel perfect display (#3469)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

Have you never found weird that you basically have to enable "Stretch to
Fit" mode in the game ?

The typical view range of (7*2+1) * 32 pixels is equal to 960 pixels.
This is almost the height of a 1080p monitor. So why use Stretch to Fit?
Where did my perfect pixel scaling go???

If you've ever had the fantasy to go pixel perfect and mess with 'Icons'
menu at top of client, you probably noticed something: the moment you
zoom through a scope or binoculars, the view range becomes 29x29 instead
of 15x15. This, unsurprisingly, doesn't work well without stretching.

This PR introduces an (optional) automatic switch of client-side zoom to
match changing view sizes. This means (by default) on 15x15 display
icons will be x2 zoom to fit a 1080p screen, and 29x29 display will
switch back to x1 zoom. This will display a small letterbox at top and
bottom of screen in 1080p but be pixel perfect. It can also be toggled
to double zoom size for bigger screens.

Because, as you might have noticed, 29 screen dimension is not double of
15 dimension, this is best used with Autofit feature, which was also
modified to work together with it.

# Explain why it's good for the game
Less blurry sprites for those willing to lose a tiny bit of screen
surface.


# Testing Photographs and Procedure
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.


![image](https://github.com/cmss13-devs/cmss13/assets/604624/a4214ebd-6c62-4d72-928e-8d394c74fd85)


![image](https://github.com/cmss13-devs/cmss13/assets/604624/039aea97-7220-4853-8f1d-6318821fac2f)


# Changelog
:cl:
add: Added optional adaptive client zooming to match in-game zoom. This
allows display to be pixel perfect on most monitors if enabled, it
removes blurring of game introduced by Stretch to Fit mode, at the cost
of slightly reduced vertical size. It is best used with Auto Fit
viewport feature enabled.
/:cl: