Grass Nerf (#3609)
# About the pull request

This PR makes it so grass (and other flora) no longer eat your bullets
when you click on them while trying to fire on a target.

# Explain why it's good for the game

The feel of your burst going into grass and the disappearing like it's a
hard object rather than... grass is awful.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
balance: Grass (and other flora) will no longer eat your bullets.
/:cl: