XO message should include code of conduct & MAIN_SHIP_NAME define usage (#3628)
# About the pull request

This PR addresses an oversight where the entry message for an XO makes
no mention of the Commander Code of Conduct, which currently has
specifics mentioned for XOs. Feel free to edit this PR to get the
message exactly how you want it. It also updates entry messages that
referred to the USS Almayer as the define MAIN_SHIP_NAME now.

# Explain why it's good for the game

XO's should be linked to information that is relevant to them.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/76988376/baf1eb66-4226-4c52-ba84-3206a39a212a)

</details>


# Changelog
:cl: Drathek
spellcheck: Added Commander Code of Conduct link to XO entry description
and fixed CO capitalization.
code: Replaced sting usage for USS Almayer to MAIN_SHIP_DEFAULT_NAME for
Souto ERT, CMB ERT, Provost ERT reporters, XO, and CO.
/:cl: